feat : add query validate middleware
