+++ src/constants/query.ts
@@ -1,0 +1,12 @@
+const MIN_LENGTH_PAGE = 1;
+const MIN_LENGTH_PAGE_NUM = 1;
+enum ORDER_BY {
+  recent = "recent",
+  oldest = "oldest",
+}
+
+export const queryLimit = {
+  MIN_LENGTH_PAGE,
+  MIN_LENGTH_PAGE_NUM,
+  ORDER_BY,
+};
