+++ src/routes/product-route.ts
@@ -1,0 +1,65 @@
+import express, { Router } from "express";
+
+import { validationHandler } from "../middlewares/error";
+import { validateAuthorization } from "../middlewares/validator-auth-input";
+import {
+  validateCreateProduct,
+  validateUpdateProduct,
+} from "../middlewares/validator-product-input";
+import { isAuthorizedProductIdParam } from "../middlewares/product";
+import { validateToken, checkToken } from "../middlewares/token";
+import productController from "../controllers/product-controller";
+
+const productRouter: Router = express.Router();
+
+productRouter
+  .post(
+    "/",
+    validateAuthorization,
+    validateCreateProduct,
+    validationHandler,
+    validateToken,
+    productController.createProduct
+  )
+  .get("/", checkToken, productController.getProductList)
+  .get(
+    "/:productId",
+    validateAuthorization,
+    validationHandler,
+    validateToken,
+    productController.getProduct
+  )
+  .patch(
+    "/:productId",
+    validateAuthorization,
+    validateUpdateProduct,
+    validationHandler,
+    validateToken,
+    isAuthorizedProductIdParam,
+    productController.modifyProduct
+  )
+  .delete(
+    "/:productId",
+    validateAuthorization,
+    validationHandler,
+    validateToken,
+    isAuthorizedProductIdParam,
+    productController.deleteProduct
+  )
+  .patch(
+    "/:productId/favorite",
+    validateAuthorization,
+    validationHandler,
+    validateToken,
+    isAuthorizedProductIdParam,
+    productController.increaseProductFavorite
+  )
+  .delete(
+    "/:productId/favorite",
+    validateAuthorization,
+    validationHandler,
+    validateToken,
+    productController.decreaseProductFavorite
+  );
+
+export default productRouter;
